ai-service/requirements.txt
torch
segment-anything
mobile-sam
open-clip-torch
fastapi
uvicorn
python-multipart
aiofiles
Pillow
numpy
